+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/Quest/QuestReward.cs
@@ -11,7 +11,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the number of experience points granted for completing the quest.
         /// </summary>
         [JsonPropertyName("experience")]
-        public long Experience { get; set; }
+        public int Experience { get; set; }
 
         /// <summary>
         /// Gets the reputation gains for completing the quest.
