+++ src/ArgentPonyWarcraftClient/Models/ProfileApi/CharacterEquipment/SelectedPower.cs
@@ -11,13 +11,13 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of the azerite power.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets the tier for the azerite power.
         /// </summary>
         [JsonPropertyName("tier")]
-        public long Tier { get; set; }
+        public int Tier { get; set; }
 
         /// <summary>
         /// Gets the spell tooltip for the azerite power.
